make path user expansion optional (#5892)
### Description

When using `LoadImage`, file paths get wrapped with `pathlib.Path` and
`.expanduser()` is called. This causes paths with a protocol prefix to
get corrupted (e.g. in my case paths starting with `s3://` are getting
one of the forward slashes removed)

This PR makes the wrapping with `Path` and call to `.expanduser()`
optional (but set to `True` by default, to preserve current behavior)

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Integration tests passed locally by running `./runtests.sh -f -u
--net --coverage`.
- [x] Quick tests passed locally by running `./runtests.sh --quick
--unittests --disttests`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated, tested `make html` command in the `docs/`
folder.

Signed-off-by: Guilherme Pires <guilherme.pires@alleninstitute.org>