5971-fix the pixelshuffle upsample shape mismatch problem. (#5982)
Signed-off-by: binliu <binliu@nvidia.com>

Fixes #5971  .

### Description
There is a mismatch shape between the conv weight and input shape in
`FlexibleUNet`, if using `pixelshuffle` upsampling method. This is
caused by the wrong initialized `pre_conv` parameter in the decoder,
which should be `default` instead of `None`. Fixed this question in this
PR by changing it to `default` and adding some test cases.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Integration tests passed locally by running `./runtests.sh -f -u
--net --coverage`.
- [ ] Quick tests passed locally by running `./runtests.sh --quick
--unittests --disttests`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated, tested `make html` command in the `docs/`
folder.

---------